+++ app/src/main/java/com/example/vnpchallenge/model/Order.java
@@ -12,12 +12,12 @@ public class Order {
     private long pickupTime;
     private String receiverName;
     private String receiverNumber;
-    private float packageWeight;
-    private float sizeW;
-    private float sizeL;
-    private float sizeH;
+    private double packageWeight;
+    private double sizeW;
+    private double sizeL;
+    private double sizeH;
     private String note;
-    private float fee;
+    private double fee;
     private int type;
 
     public Order(
@@ -26,12 +26,12 @@ public class Order {
                  long pickupTime,
                  String receiverName,
                  String receiverNumber,
-                 float packageWeight,
-                 float sizeW,
-                 float sizeL,
-                 float sizeH,
+                 double packageWeight,
+                 double sizeW,
+                 double sizeL,
+                 double sizeH,
                  String note,
-                 float fee,
+                 double fee,
                  int type) {
         this.fromAddress = fromAddress;
         this.toAddress = toAddress;
@@ -89,35 +89,35 @@ public class Order {
         this.receiverNumber = receiverNumber;
     }
 
-    public float getPackageWeight() {
+    public double getPackageWeight() {
         return packageWeight;
     }
 
-    public void setPackageWeight(float packageWeight) {
+    public void setPackageWeight(double packageWeight) {
         this.packageWeight = packageWeight;
     }
 
-    public float getSizeW() {
+    public double getSizeW() {
         return sizeW;
     }
 
-    public void setSizeW(float sizeW) {
+    public void setSizeW(double sizeW) {
         this.sizeW = sizeW;
     }
 
-    public float getSizeL() {
+    public double getSizeL() {
         return sizeL;
     }
 
-    public void setSizeL(float sizeL) {
+    public void setSizeL(double sizeL) {
         this.sizeL = sizeL;
     }
 
-    public float getSizeH() {
+    public double getSizeH() {
         return sizeH;
     }
 
-    public void setSizeH(float sizeH) {
+    public void setSizeH(double sizeH) {
         this.sizeH = sizeH;
     }
 
@@ -129,11 +129,11 @@ public class Order {
         this.note = note;
     }
 
-    public float getFee() {
+    public double getFee() {
         return fee;
     }
 
-    public void setFee(float fee) {
+    public void setFee(double fee) {
         this.fee = fee;
     }
 
@@ -162,4 +162,6 @@ public class Order {
                 ", type=" + type +
                 '}';
     }
+
+
 }
